// types/builder.ts
export interface BuilderProfile {
  wallet_address: string
  display_name: string | null
  bio: string | null
  avatar_ipfs: string | null
  twitter: string | null
  github: string | null
  telegram: string | null
  discord: string | null
  website: string | null
  created_at: string
  updated_at: string
}